requirements.txt
python-telegram-bot==20.8
python==3.10.*